0805_fix

+++ db/migrate/20200329123357_create_addresses.rb
@@ -10,7 +10,7 @@ class CreateAddresses < ActiveRecord::Migration[5.0]
       t.string :city, null: false
       t.string :street, null: false
       t.string :building
-      t.integer :telephone
+      t.bigint :telephone
       t.references :user, null: false, foreign_key: true
       t.timestamps
     end
